+++ src/renderer/effects/MenuEffect.ts
@@ -1,6 +1,0 @@
-import { createEffect } from 'effector';
-import { updateLocation } from '@renderer/stores/MenuStore.js';
-
-export const navigateTo = createEffect((destination: string) =>
-  updateLocation(destination),
-);