+++ src/renderer/stores/projects/effects/ProjectEffects.ts
@@ -1,0 +1,31 @@
+import {
+  ExportDependenciesWithNewVersionArgs,
+  ProjectCreationArgs,
+} from '@type/ProjectListenerArgs.js';
+import { createEffect } from 'effector';
+
+export const fetchProjectDetailsFx = createEffect(async (params: string) =>
+  window.projectManagement.getProjectDetails(params),
+);
+
+export const fetchProjectListFx = createEffect(() =>
+  window.projectManagement.getProjectList(),
+);
+
+export const createProjectFx = createEffect(
+  (projectCreationArgs: ProjectCreationArgs) =>
+    window.projectManagement.create(projectCreationArgs),
+);
+
+export const exportDependenciesWithNewVersionSaveDialogFx = createEffect(() =>
+  window.projectManagement.exportDependenciesWithNewVersionSaveDialog(),
+);
+
+export const exportDependenciesWithNewVersionFx = createEffect(
+  (
+    exportDependenciesWithNewVersionArgs: ExportDependenciesWithNewVersionArgs,
+  ) =>
+    window.projectManagement.exportDependenciesWithNewVersion(
+      exportDependenciesWithNewVersionArgs,
+    ),
+);